refactor: move enums to related modules
