+++ dequest/clients/_sync.py
@@ -10,9 +10,8 @@ from requests.exceptions import RequestException, Timeout
 from dequest.cache import get_cache
 from dequest.circuit_breaker import CircuitBreaker
 from dequest.config import DequestConfig
-from dequest.enums import ConsumerType
 from dequest.exceptions import CircuitBreakerOpenError, DequestError
-from dequest.http import sync_request
+from dequest.http import ConsumerType, sync_request
 from dequest.utils import extract_parameters, generate_cache_key, get_logger, map_json_to_dto, map_xml_to_dto
 
 T = TypeVar("T")
