+++ dequest/http.py
@@ -1,10 +1,25 @@
+from enum import StrEnum, auto
+
 import httpx
 import requests
 
-from dequest.enums import ConsumerType
 from dequest.utils import get_logger
 
 logger = get_logger()
+
+
+class HttpMethod(StrEnum):
+    GET = auto()
+    POST = auto()
+    PUT = auto()
+    PATCH = auto()
+    DELETE = auto()
+
+
+class ConsumerType(StrEnum):
+    XML = auto()
+    JSON = auto()
+    TEXT = auto()
 
 
 def sync_request(
